requirements.txt
streamlit
torch
transformers
torchaudio
sentencepiece
numpy
grpcio
grpcio-tools
grpcio-reflection
protobuf
soundfile